1046 - add test to return recommandations filtered by date and type
